+++ src/shapes/index.js
@@ -1,42 +1,130 @@
 import hexadecachoron from './hexadecachoron'
 import pentachoron from './pentachoron'
 import tesseract, {
-  hypercuboid,
+  hyperCuboid,
   generateTesseract,
   generateHyperCuboid,
 } from './tesseract'
-import threesphere, { generateThreeSphere } from './threesphere'
-import threetorus, { generateThreeTorus } from './threetorus'
-import flattorus, { generateFlatTorus } from './flattorus'
-import cubicpyramid, {
-  cuboidpyramid,
+import cubicPyramid, {
+  cuboidPyramid,
   generateCubicPyramid,
   generateCuboidPyramid,
-} from './cubicpyramid'
-import grandantiprism from './grandantiprism'
+} from './cubic-pyramid'
+import cubicBiPyramid, {
+  cuboidBiPyramid,
+  generateCubicBiPyramid,
+  generateCuboidBiPyramid,
+} from './cubic-bi-pyramid'
+import octahedralPyramid, {
+  octahedraloidPyramid,
+  generateOctahedralPyramid,
+  generateOctahedraloidPyramid,
+} from './octahedral-pyramid'
+import octahedralBiPyramid, {
+  octahedraloidBiPyramid,
+  generateOctahedralBiPyramid,
+  generateOctahedraloidBiPyramid,
+} from './octahedral-bi-pyramid'
+import grandAntiPrism from './grand-anti-prism'
 import hecatonicosachoron from './hecatonicosachoron'
+import hecatonicosachoronCantellated from './hecatonicosachoron-cantellated'
 import hexacosichoron from './hexacosichoron'
+import hexacosichoronBitTruncated from './hexacosichoron-bit-truncated'
+import hexacosichoronCantellated from './hexacosichoron-cantellated'
 import icositetrachoron from './icositetrachoron'
+import buckyBallPyramid from './buckyball-pyramid'
+import minkowskiSumOfTesseractAndHexadecachoron from './minkowski-sum-of-tesseract-and-hexadecachoron'
+import permutahedronA, { permutahedronAIrregular } from './permutahedron-a'
+import permutahedronB, { permutahedronBIrregular } from './permutahedron-b'
+import permutahedronC, { permutahedronCIrregular } from './permutahedron-c'
+import permutahedronD, { permutahedronDIrregular } from './permutahedron-d'
+import permutahedronF, { permutahedronFIrregular } from './permutahedron-f'
+import { /*permutahedronH,*/ permutahedronHIrregular } from './permutahedron-h'
+import crossTorus, {
+  flatTorus,
+  duoCylinder,
+  generateFlatTorus,
+  generateDuoCylinder,
+  generateUVSurface,
+} from './uv-surfaces'
+import crossHyperSurface, {
+  glome,
+  diTorus,
+  rightHyperCone,
+  obliqueHyperCone,
+  spherinder,
+  torinder,
+  cubinder,
+  tiger,
+  torisphere,
+  spheritorus,
+  generateGlome,
+  generateDiTorus,
+  generateUVWHyperSurface,
+} from './uvw-hypersurfaces'
 
 export {
   tesseract,
-  hypercuboid,
+  hyperCuboid,
   pentachoron,
   hexadecachoron,
-  threesphere,
-  threetorus,
-  flattorus,
-  cubicpyramid,
-  cuboidpyramid,
-  grandantiprism,
+  cubicPyramid,
+  cubicBiPyramid,
+  cuboidPyramid,
+  cuboidBiPyramid,
+  octahedralPyramid,
+  octahedralBiPyramid,
+  octahedraloidPyramid,
+  octahedraloidBiPyramid,
+  grandAntiPrism,
   hecatonicosachoron,
+  hecatonicosachoronCantellated,
   hexacosichoron,
+  hexacosichoronBitTruncated,
+  hexacosichoronCantellated,
   icositetrachoron,
+  buckyBallPyramid,
+  minkowskiSumOfTesseractAndHexadecachoron,
+  permutahedronA,
+  permutahedronAIrregular,
+  permutahedronB,
+  permutahedronBIrregular,
+  permutahedronC,
+  permutahedronCIrregular,
+  permutahedronD,
+  permutahedronDIrregular,
+  permutahedronF,
+  permutahedronFIrregular,
+  // permutahedronH,
+  permutahedronHIrregular,
+  crossTorus,
+  flatTorus,
+  crossHyperSurface,
+  glome,
+  diTorus,
+  rightHyperCone,
+  obliqueHyperCone,
+  spherinder,
+  torinder,
+  duoCylinder,
+  cubinder,
+  tiger,
+  torisphere,
+  spheritorus,
   generateTesseract,
   generateHyperCuboid,
-  generateCuboidPyramid,
   generateCubicPyramid,
-  generateThreeSphere,
-  generateThreeTorus,
+  generateCuboidPyramid,
+  generateCubicBiPyramid,
+  generateCuboidBiPyramid,
+  generateOctahedralPyramid,
+  generateOctahedralBiPyramid,
+  generateOctahedraloidPyramid,
+  generateOctahedraloidBiPyramid,
+  generateUVSurface,
   generateFlatTorus,
+  generateDuoCylinder,
+  generateGlome,
+  generateDiTorus,
+  generateUVWHyperSurface,
 }