+++ src/shapes/threesphere.js
@@ -1,6 +1,0 @@
-import { generateThreeTorus } from './threetorus'
-
-export const generateThreeSphere = (radius, piResolution = 8) =>
-  generateThreeTorus(radius, radius, radius, piResolution)
-
-export default generateThreeSphere(1)